requirements.txt
pandas>=2.0.0
geopandas>=0.13.0
statsmodels>=0.14.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
google-generativeai==0.5.2
PyMuPDF==1.24.4
